updated Users, inserted file and code for Thought Controller
LIke the title says, I updated the User controller, and I also insterted the controllers for both Users and THoughts. I hope that this code works. I also tried a new way of putting notes, to try and help me read my code better, so I there will be some new comments in the thought controller, dont be alarmed it is a test and so far its going well.